Add Newton CP refinement, capture analysis plots, and extracted reporting functions
- Newton-based critical point refinement (refine_to_critical_point): solves
  ∇f=0 via damped Newton's method, finds minima/maxima/saddle points unlike
  Nelder-Mead which only finds minima. Enables capture analysis for ODE objectives.

- build_known_cps_from_refinement: constructs KnownCriticalPoints from raw
  polynomial CPs by Newton-refining, deduplicating, and classifying. This is
  the key function for running capture analysis without analytically known CPs.

- Capture analysis plots (GlobtimPlots): multi-panel figure with per-type
  distance subpanels and capture rate bars. Removed redundant per-type capture
  lines panel, widened main column.

- Extracted reporting into package functions: DegreeConvergenceInfo,
  print_degree_convergence_summary, CaptureVerdict, compute_capture_verdict,
  print_capture_verdict.

- Fixed Deuflhard_4d return type annotation blocking ForwardDiff.

- Deuflhard demo: Q4 domain restriction, degrees 4-12, capture-centric flow.

+++ src/LibFunctions.jl
@@ -444,7 +444,7 @@ end
     return -0.1 * sum(5 * pi * cos(x[i]) for i in 1:4) - sum(x[i]^2 for i in 1:4)
 end
 
-@doc nothing function Deuflhard_4d(xx::AbstractVector)::Float64
+@doc nothing function Deuflhard_4d(xx::AbstractVector)
     # =======================================================
     #   Not Rescaled
     #   Domain: [-1.2, 1.2]^4.
